requirements.txt
py==1.4.31
pycodestyle==2.2.0
pytest==3.0.1
python-bitcoinrpc==1.0
simplejson==3.8.2
